Create ScopeTree data structure
A ScopeTree represents a hierarchical view of nested scopes within
a source file. The tree supports efficient insertion of scopes
(denoted by a name and a region of text within the source file),
rendering as a tree, and lookup of a region within the rendered
output containing a given point.