Add SSTException and DetailException base class errors.
SSTException derives from Exception and simply logs errors as they
are created. All other errors in the project should inherit from
SSTException.

DetailException provides a mechanism for specifying short and long
error messages. The short messages are return by repr, and may be
displayed to the user. The long messages provide more information
for the log.