Fix false positive in test_tree.Insert.test_permutations
The code intended to make a copy of the empty self.tree to use
in building the correct tree, and then make some inserts into
self.tree and compare with the copy. Instead, the correct tree
just pointed at self.tree, so that whatever happened to self.tree,
it always compared equal to the correct tree.

Changing this revealed a bug: when inserting a new scope, if the
new scope contains an existing scope, it takes the place of that
scope and adds it as a child. However, scopes before and after
the new child in the original list may also be contained by the
new scope. This is fixed by finding all of the scopes in the
original list which are contained by the new scope, removing them
from the original list, and adding them as children of the new scope.