Implement fold-on-click behavior.
When a view which was generated by SublimeScopeTree is clicked,
the smallest scope containing the location of the click is looked
up in the in-memory data structure corresponding to that view.
The region in the rendered view is then folded/unfolded. Folded
regions have only their name visible, all of their children are
hidden.